refactor(cli): lazy-load the Actions runtime so the init path stands alone
cli.ts's hoisted static imports (./index.js -> discord.js, ./db/state.js ->
better-sqlite3) evaluated before the init dispatch ran, so the wizard loaded
the full Actions stack before printing anything and a module-load failure in
that stack would crash init. The review that filed #185 assumed the
standalone repo-relay-init bin was immune — it wasn't: utils/errors.ts
imported discord.js at module top for isConfigError, and setup.ts imports
safeErrorMessage from there, so both wizard paths dragged discord.js in.

- isConfigError moves to utils/discord-errors.ts (the existing home for
  discord.js error classification); errors.ts is now dependency-free and
  documents that constraint
- cli.ts imports the runtime (RepoRelay, getChannelConfig, shouldSkipEvent,
  isConfigError) dynamically after the dispatch and env-var checks; the
  payload imports were already type-only
- new regression test registers module hooks that make discord.js and
  better-sqlite3 unresolvable and asserts both bins still reach the wizard —
  against the pre-fix dist the same hook crashes init before any output

Wizard startup drops ~150ms to ~50ms as a side effect.

Closes #185
